+++ spec/features/beis_users_can_view_other_users_spec.rb
@@ -64,5 +64,17 @@ RSpec.feature "BEIS users can can view other users" do
       expect(page).to have_content("Active? No")
       expect(page).to have_content("1 hour")
     end
+
+    scenario "an anonymised user cannot be viewed" do
+      anonymised_user = create(:inactive_user, deactivated_at: DateTime.now, anonymised_at: DateTime.now)
+
+      # Navigate from the landing page
+      visit organisation_path(user.organisation)
+      click_on(t("page_title.users.index"))
+      # Navigate to inactive users tab
+      click_on(t("tabs.users.inactive"))
+
+      expect(page).not_to have_content(anonymised_user.email)
+    end
   end
 end